+++ API/models/activity.go
@@ -6,6 +6,6 @@ type Activity struct {
 	ID    uint      `json:"id" gorm:"primary_key"`
 	Type  string    `json:"type"`
 	Unit  string    `json:"unit"`
-	Value int16     `json:"value"`
+	Value int32     `json:"value"`
 	Date  time.Time `json:"date"`
 }
